{"schema":"decodex/commit/1","summary":"Split app-server activity and dashboard scripts","authority":"manual"}
